[session] replace basic implementation with async one
